Former-commit-id: 2081122c761dba57f2e3531dadc1f936abeeda83
Former-commit-id: 4aa5dd2f6d7dd2e08faffbf4c2fca0897c4611e3

+++ registration/registration-client/src/main/java/io/mosip/registration/update/RegistrationUpdate.java
@@ -1,5 +1,8 @@
 package io.mosip.registration.update;
 
+import static io.mosip.registration.constants.RegistrationConstants.APPLICATION_ID;
+import static io.mosip.registration.constants.RegistrationConstants.APPLICATION_NAME;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -25,8 +28,12 @@ import org.springframework.stereotype.Component;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
+import io.mosip.kernel.core.logger.spi.Logger;
 import io.mosip.kernel.core.util.FileUtils;
 import io.mosip.kernel.core.util.HMACUtils;
+import io.mosip.registration.config.AppConfig;
+import io.mosip.registration.constants.LoggerConstants;
+import io.mosip.registration.controller.reg.HeaderController;
 
 /**
  * Update the Application
@@ -60,42 +67,47 @@ public class RegistrationUpdate {
 	private String mosip = "mosip";
 
 	private String versionTag = "version";
+	
+	/**o
+	 * Instance of {@link Logger}
+	 */
+	private static final Logger LOGGER = AppConfig.getLogger(RegistrationUpdate.class);
+
 
 	public boolean hasUpdate() throws IOException, ParserConfigurationException, SAXException {
-		try {
-			return !getCurrentVersion().equals(getLatestVersion());
-		} catch (RuntimeException runtimeException) {
-			return false;
-		}
+
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking for updates");
+		return !getCurrentVersion().equals(getLatestVersion());
 
 	}
 
 	private String getLatestVersion() throws IOException, ParserConfigurationException, SAXException {
-		if (latestVersion != null) {
-			return latestVersion;
-		} else {
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking for latest version started");
+		// Get latest version using meta-inf.xml
+		DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
+		DocumentBuilder db = documentBuilderFactory.newDocumentBuilder();
+		org.w3c.dom.Document metaInfXmlDocument = db.parse(new URL(serverMosipXmlFileUrl).openStream());
 
-			// Get latest version using meta-inf.xml
-			DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
-			DocumentBuilder db = documentBuilderFactory.newDocumentBuilder();
-			org.w3c.dom.Document metaInfXmlDocument = db.parse(new URL(serverMosipXmlFileUrl).openStream());
+		NodeList list = metaInfXmlDocument.getDocumentElement().getElementsByTagName(versionTag);
+		if (list != null && list.getLength() > 0) {
+			NodeList subList = list.item(0).getChildNodes();
 
-			NodeList list = metaInfXmlDocument.getDocumentElement().getElementsByTagName(versionTag);
-			if (list != null && list.getLength() > 0) {
-				NodeList subList = list.item(0).getChildNodes();
-
-				if (subList != null && subList.getLength() > 0) {
-					// Set Latest Version
-					setLatestVersion(subList.item(0).getNodeValue());
-				}
+			if (subList != null && subList.getLength() > 0) {
+				// Set Latest Version
+				setLatestVersion(subList.item(0).getNodeValue());
 			}
-
 		}
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking for latest version completed");
 		return latestVersion;
 	}
 
 	public String getCurrentVersion() throws IOException {
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking for current version started");
 		if (currentVersion != null) {
 			return currentVersion;
 		} else {
@@ -104,80 +116,89 @@ public class RegistrationUpdate {
 				setCurrentVersion((String) localManifest.getMainAttributes().get(Attributes.Name.MANIFEST_VERSION));
 			}
 		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking for current version completed");
 		return currentVersion;
 	}
 
-	public void getWithLatestJars()
-			throws IOException, ParserConfigurationException, SAXException, io.mosip.kernel.core.exception.IOException {
+	public void getWithLatestJars() throws Exception {
 
-		// Get Server Manifest
-		getServerManifest();
-
-		// Back Current Application
-		Path backUp = backUpCurrentApplication();
-
-		// replace local manifest with Server manifest
-		serverManifest.write(new FileOutputStream(new File(manifestFile)));
-
-		List<String> downloadJars = new LinkedList<>();
-		List<String> deletableJars = new LinkedList<>();
-		List<String> checkableJars = new LinkedList<>();
-
-		Map<String, Attributes> localAttributes = localManifest.getEntries();
-		Map<String, Attributes> serverAttributes = serverManifest.getEntries();
-
-		// Compare local and server Manifest
-		for (Entry<String, Attributes> jar : localAttributes.entrySet()) {
-			checkableJars.add(jar.getKey());
-			if (!serverAttributes.containsKey(jar.getKey())) {
-
-				/* unnecessary jar after update */
-				deletableJars.add(jar.getKey());
-
-			} else {
-				Attributes localAttribute = jar.getValue();
-				Attributes serverAttribute = serverAttributes.get(jar.getKey());
-				if (!localAttribute.getValue(Attributes.Name.CONTENT_TYPE)
-						.equals(serverAttribute.getValue(Attributes.Name.CONTENT_TYPE))) {
-
-					/* Jar to be downloaded */
-					downloadJars.add(jar.getKey());
-
-				}
-				serverManifest.getEntries().remove(jar.getKey());
-
-			}
-		}
-
-		for (Entry<String, Attributes> jar : serverAttributes.entrySet()) {
-			downloadJars.add(jar.getKey());
-		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Updating latest version started");
+		Path backUp = null;
 
 		try {
+			// Get Server Manifest
+			getServerManifest();
+
+			// Back Current Application
+			backUp = backUpCurrentApplication();
+			// replace local manifest with Server manifest
+			serverManifest.write(new FileOutputStream(new File(manifestFile)));
+
+			List<String> downloadJars = new LinkedList<>();
+			List<String> deletableJars = new LinkedList<>();
+			List<String> checkableJars = new LinkedList<>();
+
+			Map<String, Attributes> localAttributes = localManifest.getEntries();
+			Map<String, Attributes> serverAttributes = serverManifest.getEntries();
+
+			// Compare local and server Manifest
+			for (Entry<String, Attributes> jar : localAttributes.entrySet()) {
+				checkableJars.add(jar.getKey());
+				if (!serverAttributes.containsKey(jar.getKey())) {
+
+					/* unnecessary jar after update */
+					deletableJars.add(jar.getKey());
+
+				} else {
+					Attributes localAttribute = jar.getValue();
+					Attributes serverAttribute = serverAttributes.get(jar.getKey());
+					if (!localAttribute.getValue(Attributes.Name.CONTENT_TYPE)
+							.equals(serverAttribute.getValue(Attributes.Name.CONTENT_TYPE))) {
+
+						/* Jar to be downloaded */
+						downloadJars.add(jar.getKey());
+
+					}
+					serverManifest.getEntries().remove(jar.getKey());
+
+				}
+			}
+
+			for (Entry<String, Attributes> jar : serverAttributes.entrySet()) {
+				downloadJars.add(jar.getKey());
+			}
+
 			deleteJars(deletableJars);
 
 			// Un-Modified jars exist or not
 			checkableJars.removeAll(deletableJars);
 			checkableJars.removeAll(downloadJars);
 
+			getServerManifest();
+
 			// Download latest jars if not in local
 			checkJars(getLatestVersion(), downloadJars);
 			checkJars(getLatestVersion(), checkableJars);
 
-			setLocalManifest(getServerManifest());
+			setLocalManifest(serverManifest);
 			setServerManifest(null);
 			setLatestVersion(null);
 
-		} catch (RuntimeException exception) {
+		} catch (RuntimeException | IOException | ParserConfigurationException | SAXException exception) {
 
 			replaceBackupWithCurrentApplication(backUp);
 
 			throw exception;
 		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Updating latest version started");
 	}
 
 	private Path backUpCurrentApplication() throws IOException, io.mosip.kernel.core.exception.IOException {
-
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Backup of current version started");
 		Timestamp timestamp = new Timestamp(System.currentTimeMillis());
 		String date = timestamp.toString().replace(":", "-") + "Z";
 
@@ -204,24 +225,30 @@ public class RegistrationUpdate {
 				FileUtils.deleteDirectory(backUpFile);
 			}
 		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Backup of current version completed");
 		return backUpFolder.toPath();
 
 	}
 
 	private void replaceBackupWithCurrentApplication(Path currentApplicationbackUpPath)
 			throws io.mosip.kernel.core.exception.IOException {
-
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Replacing Backup of current version started");
 		File backUpFolder = currentApplicationbackUpPath.toFile();
 
 		FileUtils.copyDirectory(new File(backUpFolder.getAbsolutePath() + SLASH + binFolder), new File(binFolder));
 		FileUtils.copyDirectory(new File(backUpFolder.getAbsolutePath() + SLASH + libFolder), new File(libFolder));
 
 		FileUtils.copyFile(new File(backUpFolder.getAbsolutePath() + SLASH + manifestFile), new File(manifestFile));
-
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Replacing Backup of current version completed");
 	}
 
 	private void checkJars(String version, List<String> checkableJars) throws IOException {
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking of jars started");
 		for (String jarFile : checkableJars) {
 
 			String folder = jarFile.contains(mosip) ? binFolder : libFolder;
@@ -230,18 +257,29 @@ public class RegistrationUpdate {
 
 		}
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking of jars completed");
 	}
 
 	private void checkForJarFile(String version, String folderName, String jarFileName) throws IOException {
 
 		File jarInFolder = new File(folderName + jarFileName);
-		if (!jarInFolder.exists()
-				|| (!isCheckSumValid(jarInFolder, (currentVersion.equals(version)) ? localManifest : serverManifest)
-						&& FileUtils.deleteQuietly(jarInFolder))) {
 
-			// Download Jar
-			Files.copy(getInputStreamOfJar(version, jarFileName), jarInFolder.toPath());
+		// TODO Need to be removed once rxtx jars added to jfrog repo
+		List<String> notJars = java.util.Arrays.asList(
+				new String[] { "rxtxSerial.dll", "rxtxParallel.dll", "rxtxcomm-2.2", "bcprov-jdk14-138", "RXTXcomm" });
 
+		if (!notJars.contains(jarFileName)) {
+			if (!jarInFolder.exists()
+					|| (!isCheckSumValid(jarInFolder, (currentVersion.equals(version)) ? localManifest : serverManifest)
+							&& FileUtils.deleteQuietly(jarInFolder))) {
+
+				LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+						"Downloading jar : "+jarFileName+" started");
+				// Download Jar
+				Files.copy(getInputStreamOfJar(version, jarFileName), jarInFolder.toPath());
+
+			}
 		}
 
 	}
@@ -253,9 +291,13 @@ public class RegistrationUpdate {
 
 	private void deleteJars(List<String> deletableJars) throws io.mosip.kernel.core.exception.IOException {
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Deletion of jars started");
 		for (String jarName : deletableJars) {
 			deleteJar(jarName);
 		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Deletion of jars completed");
 
 	}
 
@@ -274,6 +316,8 @@ public class RegistrationUpdate {
 	}
 
 	private Manifest getLocalManifest() throws IOException {
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Geting  of local manifest started");
 		if (localManifest != null) {
 			return localManifest;
 		}
@@ -285,20 +329,22 @@ public class RegistrationUpdate {
 			setLocalManifest(new Manifest(new FileInputStream(localManifestFile)));
 
 		}
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Geting  of local manifest completed");
 		return localManifest;
 	}
 
 	private Manifest getServerManifest() throws IOException, ParserConfigurationException, SAXException {
 
-		if (serverManifest != null) {
-			return serverManifest;
-		}
-
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Geting  of server manifest started");
 		// Get latest Manifest from server
 		setServerManifest(
 				new Manifest(getInputStreamOf(serverRegClientURL + getLatestVersion() + SLASH + manifestFile)));
 		setLatestVersion(serverManifest.getMainAttributes().getValue(Attributes.Name.MANIFEST_VERSION));
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Geting  of server manifest completed");
 		return serverManifest;
 
 	}
@@ -320,11 +366,15 @@ public class RegistrationUpdate {
 	}
 
 	private boolean isCheckSumValid(File jarFile, Manifest manifest) {
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking of checksum started for jar :"+jarFile.getName());
 		String checkSum;
 		try {
 			checkSum = HMACUtils.digestAsPlainText(HMACUtils.generateHash(Files.readAllBytes(jarFile.toPath())));
 			String manifestCheckSum = (String) manifest.getEntries().get(jarFile.getName())
 					.get(Attributes.Name.CONTENT_TYPE);
+			LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+					"Checking of checksum completed for jar :"+jarFile.getName());
 			return manifestCheckSum.equals(checkSum);
 
 		} catch (IOException ioException) {
@@ -335,6 +385,8 @@ public class RegistrationUpdate {
 
 	private boolean hasSpace(int bytes) {
 
+		LOGGER.info(LoggerConstants.LOG_REG_UPDATE, APPLICATION_NAME, APPLICATION_ID,
+				"Checking of space in machine");
 		return bytes < new File("/").getFreeSpace();
 	}
